+++ components/CampusImageGallery.jsx
@@ -44,17 +44,21 @@ const GridWrapper = styled("div")`
   grid-gap: 1rem;
   grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
   grid-auto-rows: 200px;
-
-  & > div:nth-child(3) {
-    grid-row: span 2 / auto;
-  }
-
-  & > div:nth-child(5) {
-    grid-column: span 2 / auto;
-  }
-
-  & > div:nth-child(6) {
-    grid-row: span 2 / auto;
+  @media (min-width: 1165px) {
+    & > div:nth-child(3) {
+      grid-row: span 2 / auto;
+    }
+  
+    & > div:nth-child(5) {
+      grid-column: span 2 / auto;
+    }
+  
+    & > div:nth-child(6) {
+      grid-row: span 2 / auto;
+    }
+    & > div:nth-child(10) {
+        grid-column: span 2 / auto;
+    }
   }
 `;
